Fix: move get chat stream to server/utils folder
