+++ server/api/email-followup.ts
@@ -1,9 +1,6 @@
 import chalk from 'chalk'
-import { getChatStream } from '~/utils'
 
 export default defineEventHandler(async (event) => {
-  const agentRole =
-    'You are a customer service representative for a High Park Livery, a luxury transportation company. '
   const openai = event.context.openai
   const body = await readBody(event)
   console.log(chalk.green(`[EVENT_BODY]`), body)
